pyproject.toml
[project]
name = "alaye-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nuitka>=2.8.9",
    "python-dotenv>=1.2.1",
    "pywebview>=6.1",
    "PyQt6>=6.5.0; sys_platform == 'linux'",
    "PyQt6-WebEngine>=6.5.0; sys_platform == 'linux'",
    "qtpy>=2.4.0; sys_platform == 'linux'",
    "pexpect>=4.9.0; sys_platform != 'win32'",
    "pywinpty>=2.0.12; sys_platform == 'win32'",
    "pydantic>=2.0.0",
    "fastapi>=0.124.4",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.45",
    "alembic>=1.12.0",
    "aiosqlite>=0.19.0",
    "greenlet>=3.3.0",
]

[dependency-groups]
test = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]
dev = [
    "autoflake>=2.3.0",
    "black>=25.12.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "flake8-docstrings>=1.7.0",
    "flake8-import-order>=0.18.2",
    "pre-commit>=3.6.0",
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | frontend
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]
known_third_party = ["pytest", "pydantic"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
combine_as_imports = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
